update: Implement requested changes
